+++ front-end/src/pages/professor/course/classroom/ProfessorClassroom.module.css
@@ -4,120 +4,3 @@
   user-select: none;
   background-color: black;
 }
-.mainContainer {
-  display: flex;
-  flex-direction: column;
-  width: fit-content;
-}
-
-.mainHeader {
-  width: 100%;
-  height: 155px;
-  display: flex;
-  align-items: center;
-  padding: 0 24px 0 48px;
-  justify-content: space-between;
-}
-
-.accessCodeContainer {
-  display: flex;
-  gap: 5px;
-}
-
-.accessCode {
-  width: 206px;
-  height: 48px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  color: white;
-  background-color: var(--gray-800);
-  border-radius: 12px;
-  font: var(--web-title5-medium);
-}
-
-.expandCode {
-  width: 48px;
-  height: 48px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  border: 2px solid rgba(255, 255, 255, 0.5);
-  border-radius: 10px;
-}
-
-.expandSvg {
-  width: 23px;
-  height: 23px;
-  color: var(--gray-300);
-}
-
-.buttonContainer {
-  display: flex;
-  align-items: center;
-  gap: 11px;
-}
-
-.noPdfContainer {
-  width: 1370px;
-  height: 770px;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-}
-
-.noPdfDesc {
-  font: var(--web-header1-bold);
-  color: white;
-  margin-bottom: 38px;
-}
-
-.mainFooter {
-  width: 100%;
-  height: 155px;
-  display: flex;
-  align-items: center;
-  padding: 0 24px 0 48px;
-  justify-content: space-between;
-}
-
-.fileButton {
-  width: 201px;
-  height: 61px;
-  border: 3px solid rgba(255, 255, 255, 0.5);
-  border-radius: 12px;
-  color: white;
-  font: var(--web-button3-bold);
-}
-
-.sideBarContainer {
-  display: flex;
-  width: 100%;
-  flex-direction: column;
-}
-.sideContentContainer {
-  display: flex;
-  width: 100%;
-  flex-direction: column;
-  padding: 18px 20px 0 20px;
-  background-color: var(--black);
-}
-
-.requestsContainer {
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  margin-top: 92px;
-  margin-bottom: 36px;
-}
-
-.closeButton {
-  width: 100%;
-  height: 100%;
-  font: var(--web-button2-bold);
-  color: white;
-  background-color: var(--blue-400);
-  text-align: center;
-}
