+++ front-end/src/pages/professor/course/classroom/components/PDFSidebar.module.css
@@ -1,0 +1,30 @@
+.sideBarContainer {
+  display: flex;
+  width: 100%;
+  flex-direction: column;
+}
+.sideContentContainer {
+  display: flex;
+  width: 100%;
+  flex-direction: column;
+  padding: 18px 20px 0 20px;
+  background-color: var(--black);
+}
+
+.requestsContainer {
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  margin-top: 92px;
+  margin-bottom: 36px;
+}
+
+.closeButton {
+  width: 100%;
+  height: 100%;
+  font: var(--web-button2-bold);
+  color: white;
+  background-color: var(--blue-400);
+  text-align: center;
+}
